Constants done, Drawable added, next is camera

+++ VeritasCpp/Engine/Arch/InputLayout.cpp
@@ -3,7 +3,7 @@
 InputLayout::InputLayout(VeritasEngine& gfx,
 	std::span<VINPUT_ELEMENT> layout)
 {
-	GetDevice(gfx)->CreateInputLayout(layout.data(), layout.size(), nullptr, 0, &pInputLayout);
+	GetDevice(gfx)->CreateInputLayout(layout.data(), (uint32_t)layout.size(), nullptr, 0, &pInputLayout);
 }
 
 void InputLayout::Bind(VeritasEngine& gfx) noexcept
